% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_hits.R
\name{filter_hits}
\alias{filter_hits}
\title{A function that filters out false positive sequence scan hits once threshold is provided.}
\usage{
filter_hits(all_hits, threshold)
}
\arguments{
\item{all_hits}{A dataframe containint all hits reported by sequence scan}

\item{threshold}{A threshold of combined score of location and sequence similarity that is used to filter out false positive scan hits.}
}
\value{
a dataframe with hits that are considered to be true positive sequence scan hits
}
\description{
A function that filters out false positive sequence scan hits once threshold is provided.
}
