% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_threshold.R
\name{find_threshold}
\alias{find_threshold}
\title{A function that creates normalises score and location position
from sequence scann hits and suggests an optimal threshold for filtering
false positive hits.
To filter false positives we have implemented location constraints of core promoter TFs
To use default constraints of these PWMs, use "default" parameter for constraints parameter.
User can also submit custom constraints too.}
\usage{
find_threshold(all_hits, constraints = "default")
}
\arguments{
\item{all_hits}{A dataframe containint all hits reported by sequence scan}

\item{constraints}{location constraints where PWM can be found. User can use default
defined location constraints, or define custom constraints in form of a numeric vector with start and end coordinate.}
}
\value{
suggested threshold value for filtering out false positive hits
}
\description{
A function that creates normalises score and location position
from sequence scann hits and suggests an optimal threshold for filtering
false positive hits.
To filter false positives we have implemented location constraints of core promoter TFs
To use default constraints of these PWMs, use "default" parameter for constraints parameter.
User can also submit custom constraints too.
}
