% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CAGE_promoters.R
\name{get_CAGE_promoters}
\alias{get_CAGE_promoters}
\title{Create promoters centered on dominant TSS CAGE tag clusters with annotations from Biomart promoters.}
\usage{
get_CAGE_promoters(tc_gr, promoters, keep_all_promoters = FALSE,
  upstream = 500, downstream = 500, protein_coding_only = TRUE)
}
\arguments{
\item{tc_gr}{GenomicRanges object containing tag clusters extracted from CAGEset}

\item{promoters}{GenomicRanges object containing promoters extracted from Biomart}

\item{keep_all_promoters}{logical: in case there are multiple TC assigned to a promoter,
do you want to keep all of them or just the one with highest expression}

\item{upstream}{upstream region from TSS to be included in promoter region}

\item{downstream}{downstream region from TSS to be included in promoter region}

\item{protein_coding_only}{logical: will returned promoters be promoter of
             only promoter coding genes

This function creates promoter ranges from CAGE tag clusters.}
}
\value{
a promoter regions centered on dominant TSS extracted from CAGE tag clusters
}
\description{
Create promoters centered on dominant TSS CAGE tag clusters with annotations from Biomart promoters.
}
\keyword{promoters}
