% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiple_POs.R
\name{plot_multiple_POs}
\alias{plot_multiple_POs}
\title{Plot mutliple PO results in one barchart}
\usage{
plot_multiple_POs(list_POs, use_cage = TRUE)
}
\arguments{
\item{list_POs}{a list object where each sublist is a result of promoter_onto function}

\item{use_cage}{logical specifying whether to plot}
}
\value{
plot object to which you can add extra ggplot parameters
}
\description{
This function allows to plot results of promoter ontology for multiple groups of genes in one plot, making comparison between different gene groups easy. Function returns ggplot2 object, so in the case additional customisation of plot is needed, extra parameters can be added additionaly with "+" sign.
}
