% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promoter_onto.R
\name{promoter_onto}
\alias{promoter_onto}
\title{Overenrichment analysis of promoter features for a group of genes}
\usage{
promoter_onto(promoter_matrix, sample, proms, features = c(1:11, 16:17),
  save_plot = TRUE, label, all = FALSE)
}
\arguments{
\item{promoter_matrix}{promoter matrix that for each promoter (row), describes its core promoter features (columns)}

\item{sample}{gene names for which we will calculate overrepresentation of features}

\item{proms}{genomic ranges object containing promoters of all active genes centered on dominant TSS}

\item{features}{which features to test, provided as vector of column indices of promoter_matrix matrix}

\item{save_plot}{should results be saved, or just presented in console, if yes, file is going to be saved as PDF to working directory}

\item{label}{filename for the plot}

\item{all}{weather to report stats of all features (TRUE), or only significan ones (FALSE)}
}
\value{
a list with significance features of overrepresentation test
}
\description{
Promoter_onto performs oveerenrichment analysis of promoter features for genes in sample compared to the background. Promoter_matrix object contains annotations of promoter features for a background set. Overenrichment analysis also looks at CAGE features like tpm expression and promoter interquantile width, for that reason promoter input is needed too. By default this function plots overenrichment plots as a pdf to the file in current directory.
}
\keyword{overrepresentation}
